+++ src/arkhados/actions/castspellactions/CastOnGroundAction.java
@@ -51,8 +51,8 @@ public class CastOnGroundAction extends EntityAction {
     public boolean update(float tpf) {
         final SpellCastControl castControl = super.spatial.getControl(SpellCastControl.class);
         final Vector3f adjustedTarget = castControl.getClosestPointToTarget(this.spell).setY(0.1f);
-        final Long playerId = super.spatial.getUserData(UserDataStrings.PLAYER_ID);
-        final Long entityId = worldManager.addNewEntity(spell.getName(), adjustedTarget, Quaternion.IDENTITY, playerId);
+        final Integer playerId = super.spatial.getUserData(UserDataStrings.PLAYER_ID);
+        final int entityId = worldManager.addNewEntity(spell.getName(), adjustedTarget, Quaternion.IDENTITY, playerId);
 
         final Spatial entity = worldManager.getEntity(entityId);
         final AreaEffectControl aoeControl = entity.getControl(AreaEffectControl.class);
